Sistema Refatorado
Efetuado refatoração do sistema utilizando conceitos de SOLID e Strategy Pattern
