+++ TheatricalPlayersRefactoringKata.Tests/StatementPrinterTests.cs
@@ -2,6 +2,8 @@ using System;
 using System.Collections.Generic;
 using ApprovalTests;
 using ApprovalTests.Reporters;
+using TheatricalPlayersRefactoringKata.Core;
+
 using Xunit;
 
 namespace TheatricalPlayersRefactoringKata.Tests;
