+++ TheatricalPlayersRefactoringKata/Models/Performance.cs
@@ -1,4 +1,4 @@
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Core;
 
 public class Performance
 {
