+++ TheatricalPlayersRefactoringKata/Models/Play.cs
@@ -1,4 +1,4 @@
-namespace TheatricalPlayersRefactoringKata;
+namespace TheatricalPlayersRefactoringKata.Core;
 
 public class Play
 {
@@ -10,11 +10,11 @@ public class Play
     public int Lines { get => _lines; set => _lines = value; }
     public string Type { get => _type; set => _type = value; }
 
-    public Play(string name, int lines, string type) {
+    public Play(string name, int lines, string type)
+    {
         _name = name;
         _lines = lines;
         _type = type;
     }
 }
 
-
